[refactor] ItemController, ItemService conflict merge
